wire it up
