+++ js/examples/yjs/package.json
@@ -32,7 +32,9 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "uuid": "^9.0.0",
-    "ws": "^8.11.0",
+    "ws": "^8.11.0"
+  },
+  "peerDependencies": {
     "yjs": "^13.5.43"
   },
   "devDependencies": {
