+++ js/examples/yjs/src/App.tsx
@@ -1,25 +1,17 @@
 import {
+  $createParagraphNode,
+  $createTextNode,
   $getRoot,
-  $getSelection,
-  BLUR_COMMAND,
-  COMMAND_PRIORITY_LOW,
-  EditorState,
-  FOCUS_COMMAND,
+  LexicalEditor,
 } from "lexical";
 
 import { LexicalComposer } from "@lexical/react/LexicalComposer";
 import { RichTextPlugin } from "@lexical/react/LexicalRichTextPlugin";
 import { ContentEditable } from "@lexical/react/LexicalContentEditable";
 import { HistoryPlugin } from "@lexical/react/LexicalHistoryPlugin";
-import { OnChangePlugin } from "@lexical/react/LexicalOnChangePlugin";
-import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
 import { CollaborationPlugin } from "@lexical/react/LexicalCollaborationPlugin";
 import LexicalErrorBoundary from "@lexical/react/LexicalErrorBoundary";
-import {
-  $convertFromMarkdownString,
-  $convertToMarkdownString,
-  TRANSFORMERS,
-} from "@lexical/markdown";
+import { TRANSFORMERS } from "@lexical/markdown";
 import { HeadingNode, QuoteNode } from "@lexical/rich-text";
 import { TableCellNode, TableNode, TableRowNode } from "@lexical/table";
 import { ListItemNode, ListNode } from "@lexical/list";
@@ -80,26 +72,28 @@ export default function App({ ctx }: { ctx: Ctx }) {
           const doc = new Y.Doc();
           yjsDocMap.set(id, doc);
 
-          // TODO: we'll need to make this conform to the websocket provider
-          // interface -- or modify lexical.
-          // Seems best to just pretend we're a websocket provider for now.
-          const crsqlYjs = await CrsqlYjs(ctx.db, ctx.rx, id, doc);
-
-          const provider = new WebsocketProvider(
-            "ws://localhost:1234",
-            id,
-            doc
-          );
-
-          return provider;
+          return new CrsqlYjs({
+            db: ctx.db,
+            rx: ctx.rx,
+            docid: id,
+            doc,
+          });
         }}
         // Optional initial editor state in case collaborative Y.Doc won't
         // have any existing data on server. Then it'll user this value to populate editor.
         // It accepts same type of values as LexicalComposer editorState
         // prop (json string, state object, or a function)
-        // initialEditorState={initialEditorState}
+        initialEditorState={initialEditorState}
         shouldBootstrap={true}
       />
     </LexicalComposer>
   );
+}
+
+function initialEditorState(editor: LexicalEditor): void {
+  const root = $getRoot();
+  const paragraph = $createParagraphNode();
+  const text = $createTextNode("Welcome to collab!");
+  paragraph.append(text);
+  root.append(paragraph);
 }
