+++ js/yjs-provider/package.json
@@ -24,7 +24,9 @@
     "@types/throttle-debounce": "^5.0.0",
     "@vlcn.io/rx-tbl": "workspace:*",
     "@vlcn.io/xplat-api": "workspace:*",
-    "throttle-debounce": "^5.0.0",
+    "throttle-debounce": "^5.0.0"
+  },
+  "peerDependencies": {
     "yjs": "^13.5.43"
   }
 }
